+++ bidsmreye/config/default_filter_file.json
@@ -14,7 +14,6 @@
         "extension": "npz"
     },
     "eyetrack": {
-        "desc": "bidsmreye",
         "suffix": "^eyetrack$$",
         "extension": "tsv"
     }
