+++ bidsmreye/generalize.py
@@ -16,18 +16,19 @@ from deepmreye.util import data_generator
 from deepmreye.util import model_opts
 from rich import print
 
-from bidsmreye.quality_control import quality_control
+from bidsmreye.bids_utils import check_layout
+from bidsmreye.bids_utils import create_bidsname
+from bidsmreye.bids_utils import get_dataset_layout
+from bidsmreye.bids_utils import list_subjects
+from bidsmreye.configuration import Config
+from bidsmreye.logging import bidsmreye_log
+from bidsmreye.quality_control import quality_control_output
 from bidsmreye.utils import add_sidecar_in_root
-from bidsmreye.utils import check_layout
-from bidsmreye.utils import Config
-from bidsmreye.utils import create_bidsname
 from bidsmreye.utils import create_dir_for_file
-from bidsmreye.utils import get_dataset_layout
-from bidsmreye.utils import list_subjects
 from bidsmreye.utils import move_file
 from bidsmreye.utils import set_this_filter
 
-log = logging.getLogger("bidsmreye")
+log = bidsmreye_log(name="bidsmreye")
 
 
 def create_and_save_figure(
@@ -210,7 +211,7 @@ def generalize(cfg: Config) -> None:
     log.info("GENERALIZING")
     log.info(f"Using model: {cfg.model_weights_file}")
 
-    layout_out = get_dataset_layout(cfg.output_folder)
+    layout_out = get_dataset_layout(cfg.output_dir)
     check_layout(cfg, layout_out)
 
     add_sidecar_in_root(layout_out)
@@ -221,4 +222,4 @@ def generalize(cfg: Config) -> None:
 
         process_subject(cfg, layout_out, subject_label)
 
-    quality_control(cfg)
+    quality_control_output(cfg)
