+++ tests/data/.gitignore
@@ -1,1 +1,2 @@
 moae_fmriprep
+ds000201-der/derivatives
