+++ tests/test_prepare.py
@@ -2,18 +2,18 @@ from __future__ import annotations
 
 from pathlib import Path
 
+from bidsmreye.bids_utils import get_dataset_layout
 from bidsmreye.prepare_data import combine_data_with_empty_labels
-from bidsmreye.utils import get_dataset_layout
 
 
 def test_combine_data_with_empty_labels():
 
-    output_location = Path().resolve()
-    output_location = output_location.joinpath("tests", "data", "bidsmreye")
+    output_dir = Path().resolve()
+    output_dir = output_dir.joinpath("tests", "data", "bidsmreye")
 
-    layout_out = get_dataset_layout(output_location)
+    layout_out = get_dataset_layout(output_dir)
 
-    file = output_location.joinpath(
+    file = output_dir.joinpath(
         "sub-01",
         "func",
         "sub-01_task-nback_run-01_space-MNI152NLin2009cAsym_desc-eye_mask.p",
@@ -23,7 +23,7 @@ def test_combine_data_with_empty_labels():
 
     assert no_label_file.is_file()
 
-    expected_file = output_location.joinpath(
+    expected_file = output_dir.joinpath(
         "sub-01",
         "func",
         "sub-01_task-nback_run-01_space-MNI152NLin2009cAsym_desc-nolabel_bidsmreye.npz",
