+++ tests/test_quality_control.py
@@ -10,15 +10,16 @@ from .utils import create_basic_data
 from .utils import create_basic_json
 from .utils import create_confounds_tsv
 from .utils import return_bidsmreye_eyetrack_tsv
+from bidsmreye.bids_utils import create_bidsname
+from bidsmreye.bids_utils import get_dataset_layout
+from bidsmreye.configuration import Config
 from bidsmreye.quality_control import add_qc_to_sidecar
 from bidsmreye.quality_control import compute_displacement_and_outliers
 from bidsmreye.quality_control import compute_robust_outliers
 from bidsmreye.quality_control import get_sampling_frequency
 from bidsmreye.quality_control import perform_quality_control
-from bidsmreye.quality_control import quality_control
-from bidsmreye.utils import Config
-from bidsmreye.utils import create_bidsname
-from bidsmreye.utils import get_dataset_layout
+from bidsmreye.quality_control import quality_control_input
+from bidsmreye.quality_control import quality_control_output
 
 
 def time_series():
@@ -117,12 +118,12 @@ def test_compute_robust_with_nan():
     assert outliers == expected_outliers
 
 
-def test_quality_control():
+def test_quality_control_output():
 
     create_basic_json()
 
-    output_location = Path().resolve()
-    output_location = output_location.joinpath("tests", "data")
+    output_dir = Path().resolve()
+    output_dir = output_dir.joinpath("tests", "data")
 
     confounds_tsv = return_bidsmreye_eyetrack_tsv()
 
@@ -130,26 +131,56 @@ def test_quality_control():
     df.to_csv(confounds_tsv, sep="\t", index=False)
 
     cfg = Config(
-        output_location,
-        output_location,
+        output_dir,
+        output_dir,
     )
 
-    quality_control(cfg)
+    quality_control_output(cfg)
+
+
+def test_quality_control_input():
+
+    input_dir = Path().resolve().joinpath("tests", "data", "ds000201-der")
+    output_dir = input_dir.joinpath("derivatives", "bidsmreye")
+
+    cfg = Config(
+        input_dir,
+        output_dir,
+    )
+
+    quality_control_input(cfg)
 
 
 def test_perform_quality_control():
 
     create_basic_json()
 
-    output_location = Path().resolve()
-    output_location = output_location.joinpath("tests", "data", "bidsmreye")
-    layout_out = get_dataset_layout(output_location)
+    output_dir = Path().resolve()
+    output_dir = output_dir.joinpath("tests", "data", "bidsmreye")
+    layout = get_dataset_layout(output_dir)
     confounds_tsv = return_bidsmreye_eyetrack_tsv()
 
     df = pd.DataFrame(create_basic_data())
     df.to_csv(confounds_tsv, sep="\t", index=False)
 
-    perform_quality_control(layout_out, confounds_tsv)
+    perform_quality_control(layout, confounds_tsv)
+
+
+def test_perform_quality_control_with_different_output():
+
+    input_dir = Path().resolve().joinpath("tests", "data", "ds000201-der")
+    layout_in = get_dataset_layout(input_dir)
+
+    output_dir = input_dir.joinpath("derivatives", "bidsmreye")
+    layout_out = get_dataset_layout(output_dir)
+
+    confounds_tsv = layout_in.get(
+        return_type="filename", subject="9001", suffix="eyetrack", extension=".tsv"
+    )[0]
+
+    perform_quality_control(
+        layout_in=layout_in, confounds_tsv=confounds_tsv, layout_out=layout_out
+    )
 
 
 def test_add_qc_to_sidecar():
@@ -158,9 +189,9 @@ def test_add_qc_to_sidecar():
 
     create_confounds_tsv()
 
-    output_location = Path().resolve()
-    output_location = output_location.joinpath("tests", "data", "bidsmreye")
-    layout_out = get_dataset_layout(output_location)
+    output_dir = Path().resolve()
+    output_dir = output_dir.joinpath("tests", "data", "bidsmreye")
+    layout_out = get_dataset_layout(output_dir)
 
     confounds_tsv = return_bidsmreye_eyetrack_tsv()
     confounds = pd.read_csv(confounds_tsv, sep="\t")
